hms/requirements.txt
# Core
Django==5.0.6
djangorestframework==3.15.2
drf-nested-routers==0.94.1
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# Environment & config
python-decouple==3.8
dj-database-url==2.2.0

# Security & encryption
argon2-cffi==23.1.0
django-cors-headers==4.4.0

# Filtering & search
django-filter==24.2

# Storage (S3)
boto3==1.34.144
django-storages==1.14.4

# Async tasks
celery==5.4.0
redis==5.0.7
django-celery-beat==2.6.0
django-celery-results==2.5.1

# Monitoring & logging
sentry-sdk==2.10.0

# API documentation
drf-spectacular==0.27.2

# Utilities
Pillow==10.4.0
python-dateutil==2.9.0
phonenumbers==8.13.42

# Dev / testing only
pytest==8.3.2
pytest-django==4.8.0
factory-boy==3.3.1
coverage==7.6.0
black==24.4.2
flake8==7.1.0
